rag_pipeline/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
oracledb
oci
pydantic
pydantic-settings
openpyxl
pandas
python-pptx
pypdf
PyMuPDF
pytest
python-multipart
